requirements.txt
streamlit
pandas
scikit-learn
nltk
google-genai